requirements.txt
Django==4.2.7
asgiref
sqlparse
tzdata
django-simple-captcha